Add tests on MemberServiceImpl

+++ src/main/java/org/example/testtasknerdy/exception/MemberHasBorrowedBooksException.java
@@ -3,7 +3,7 @@ package org.example.testtasknerdy.exception;
 public class MemberHasBorrowedBooksException extends RuntimeException {
     private static final String MESSAGE = "Unable deletion. Member with id = %d has borrowed books.";
 
-    public MemberHasBorrowedBooksException(final int id) {
+    public MemberHasBorrowedBooksException(final long id) {
         super(String.format(MESSAGE, id));
     }
 }
